PlayerCharacter Iteration One
Added ParentStation and InteractableInterface, which will be used later in development.
Added MainMenuCharacter and Data assets, used in the Main Menu level for choosing your level.  Also started design of Drive Through area in Main Menu.
Added ItemData, a library used to store all structs and enums surrounding item data.
Added ParentItem, which contains a static mesh component and a function for setting up the Item.
Added PlayerCharacter, who can move, look around and pick up items.

Iteraion One of the character features two hand objects - LeftHand and RightHand - that can be activated by pressing or holding left or right mouse button respectively.  While activated, overlapping with an item picks it up, attaching it to the hand that collided with it.
While I liked how this worked, it was hard to aim each hand.  In the next iteration, I will be implementing the ability to see what you are picking up before activating the hands.