Plating
Modified Crafting slightly, now will craft from child classes of AParentItem, without worrying about the item order.
 - Still need a little change in spawning other classes though
Implemented Plate class.  Added bool bDirty, will be used later
Modified PlaceItem trace to check if the hit item is a plate.  If it is, it attaches the item directly to the plate instead of placing it.  Items can still be removed from the plate as well.