Stackables
Moved AttachedItems from AParentItem to APlate.  This was only used by plate as Stackables use the StackedItem TSet.
Added StackedItem TSet to AParentItem
 - A Set is used as it keeps attached items in order of attachment.  We iterate backwards across the set when detaching.
Implement DetachStack function in AParentItem, used when collecting an item from a stack.
 - As an example, you have a stack of bun bottom, burger, lettuce leaf and bun top.  Removing the burger will detach the bun top and lettuce leaf.  If you collected the leaf instead, only the bun top would detach.
Modified APlayerCharacter Place Trace to check what the class of the AttachedTo item is.