Single Item Inventory
Deprecated ChangeItem and TogglePlaceMode functions
Deprecated CurrentHeldItem
Changed HeldItems TArray to single HeldItem pointer
Changed comparisons for bPlaceMode and HeldItem.Num to check if HeldItem is nullptr or not

Changed having a multi slot item 'inventory' to a single item one for simplification.  Players are not really ever going to pick up multiple items at once.  This also allows for placing with a single button without toggling between modes and reduces tick count when in range with a pickup while holding an item.