Taking from Containers
Marked bStoreable for Deprecation, actually don't need it as Containers have an array of accepted items already
When placing an item in a container, HeldItem now becomes nullptr after Destroy
Removed ContainerRange and Overlap events from Container, uneeded
RemoveItemFromContainer now checks in Amount is 0 after removing an item.  If it is 0, the ContainedItem struct in container is emptied.