Comparing Plate to Order
Added FStackItem struct, which is used to define an item in an order stack and the rule on how it can be placed in that stack
 - For example, a Burger Bun Top with a rule of MustBeFirst can only be ontop of the stack, while a Cheese slice with the rule CantBeBounds makes it so it can be anywhere except the top and bottom
Added FStackAsID, which contains an array of item IDs, converted from a plate
Added FOrderable struct, which is used in a data table to state every item that an NPC can order
Added FOrder, which is just all of one NPC's orderables placed in one struct for storage
Added AimPosition to PlayerHUD, removed drawing a debug line
Added pointer to Orderables data table in OrderManager, as well as an array of all Orderables that can be ordered at that time.
Added Money properties to OrderManager.
Started implementing CompareOrder function.
 - Currently checks if the first item of an orderable is contained in each stack of InStacks.  If there are more than 1 stacks with that item, then checks the second item.  This repeats until all items are checked.
 - If the check ever exceeds the amount of items in the orderables, then there must be duplicate items, so it takes the first item.
 - If no match is found, the compare is cancelled